删除默认最小金额小数点

+++ app/src/main/java/org/cxct/sportlottery/ui/betList/holder/BetInfoItemViewHolder.kt
@@ -209,7 +209,7 @@ class BetInfoItemViewHolder(
         etBet.apply {
             if (itemData.input == null) {
                 val minBet = itemData.parlayOdds?.min ?: 0
-                itemData.input = minBet.toDouble().toString()
+                itemData.input = minBet.toString()
             }
             itemData.inputBetAmountStr = itemData.input
             itemData.betAmount = itemData.input!!.toDouble()
